pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api_exchange_core"
version = "0.1.0"
description = "Core framework for building serverless data exchange and integration pipelines"
readme = "README.md"
authors = [
    {name = "API Exchange Team", email = "info@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "azure-storage-queue>=12.6.0",
    "azure-functions>=1.14.0",
    "azure-servicebus>=7.11.0",
    "psycopg2-binary>=2.9.7",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.8.0",
    "alembic>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=4.0.0",
    "pytest-bdd>=8.0.0",
    "factory-boy>=3.3.0",
    "requests-mock>=1.12.0",
    "black>=23.0.0", 
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "bandit>=1.8.0",
    "autoflake>=2.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "matplotlib>=3.10.0",
    "plotly>=6.0.0",
    "networkx>=3.4.0",
    "prometheus-client>=0.21.0",
]
e2e = [
    "azure-functions>=1.18.0",
    "azure-storage-queue>=12.8.0", 
    "azure-storage-blob>=12.19.0",
    "requests>=2.31.0",
    "python-json-logger>=2.0.7",
]

[tool.setuptools]
packages = {find = {exclude = ["tests*"]}}

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"